+++ src/core/Rectangle.java
@@ -8,8 +8,8 @@ package core;
  */
 public class Rectangle {
 
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 	private Size size;
 
 	/**
@@ -24,7 +24,7 @@ public class Rectangle {
 	 * @param height
 	 *            The height of the rectangle.
 	 */
-	public Rectangle(int x, int y, int width, int height) {
+	public Rectangle(float x, float y, float width, float height) {
 		this.x = x;
 		this.y = y;
 		this.size = new Size(width, height);
@@ -65,7 +65,7 @@ public class Rectangle {
 	 * @param y
 	 *            The y coordinate.
 	 */
-	public void moveTo(int x, int y) {
+	public void moveTo(float x, float y) {
 		this.setPosition(x, y);
 	}
 	/**
@@ -79,7 +79,7 @@ public class Rectangle {
 	 * @param y
 	 *            The y coordinate.
 	 */
-	public void moveBy(int x, int y) {
+	public void moveBy(float x, float y) {
 		this.setPosition(this.getX() + x, this.getY() + y);
 	}
 
@@ -90,7 +90,7 @@ public class Rectangle {
 	 * @param x
 	 *            The x coordinate.
 	 */
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 	/**
@@ -99,7 +99,7 @@ public class Rectangle {
 	 * @param y
 	 *            The y coordinate.
 	 */
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 	/**
@@ -108,7 +108,7 @@ public class Rectangle {
 	 * @param width
 	 *            The width.
 	 */
-	public void setWidth(int width) {
+	public void setWidth(float width) {
 		this.size.setWidth(width);
 	}
 	/**
@@ -117,7 +117,7 @@ public class Rectangle {
 	 * @param height
 	 *            The height.
 	 */
-	public void setHeight(int height) {
+	public void setHeight(float height) {
 		this.size.setHeight(height);
 	}
 	/**
@@ -128,7 +128,7 @@ public class Rectangle {
 	 * @param y
 	 *            The y coordinate.
 	 */
-	public void setPosition(int x, int y) {
+	public void setPosition(float x, float y) {
 		this.setX(x);
 		this.setY(y);
 	}
@@ -140,7 +140,7 @@ public class Rectangle {
 	 * @param height
 	 *            The height.
 	 */
-	public void setDimension(int width, int height) {
+	public void setDimension(float width, float height) {
 		this.setWidth(width);
 		this.setHeight(height);
 	}
@@ -151,7 +151,7 @@ public class Rectangle {
 	 * 
 	 * @return The x coordinate.
 	 */
-	public int getX() {
+	public float getX() {
 		return this.x;
 	}
 	/**
@@ -159,7 +159,7 @@ public class Rectangle {
 	 * 
 	 * @return The y coordinate.
 	 */
-	public int getY() {
+	public float getY() {
 		return this.y;
 	}
 	/**
@@ -167,7 +167,7 @@ public class Rectangle {
 	 * 
 	 * @return The width.
 	 */
-	public int getWidth() {
+	public float getWidth() {
 		return this.size.getWidth();
 	}
 	/**
@@ -175,7 +175,7 @@ public class Rectangle {
 	 * 
 	 * @return The height.
 	 */
-	public int getHeight() {
+	public float getHeight() {
 		return this.size.getHeight();
 	}
