+++ src/tilemap/Tile.java
@@ -94,7 +94,7 @@ public class Tile implements JSONSerializable, GLRenderableObject, Comparable {
 	// Interface methods
 	@Override
 	public void render(Rectangle object, Rectangle target,
-			GLAutoDrawable canvas, int zIndex) {
+			GLAutoDrawable canvas, float zIndex) {
 
 		if (this.sprite != null)
 			this.sprite.render(object, target, canvas, 0);
